Add syntax highlighting CSS to blog posts
Jekyll automatically uses the `rouge` library to perform syntax
highlighting on fenced code blocks. When building the site, Github Pages
makes sure the same highlighting happens.

This causes the code in the blog posts to be annotated with differnt
syntactical elements. All that's required, then, is to create a
stylesheet that targets these elements and use that stylesheet within
the blog posts.

The stylesheet comes from [jekyll-pigments-themes](https://github.com/jwarby/jekyll-pygments-themes)
from jwarby, which is licensed under the Unlicense (as close to public
domain as possible).