Blog post: A graphical introduction to dynamic programming
This post requires math rendering, so this commit introduces the
infrastructure to render math using the MathJax library. This library
performs the rendering on the client side, which is not necessarily
something I'm happy with. However, with the default Jekyll setup
available in Github Pages, this is the best I can do.