Add RSS feed
Github pages supports the use of the jekyll-feed plugin for Jekyll, so
I'm using that. To make the feed a little better, I've added a title and
description to the global config.

For the same reason, I've added my global author information. In doing
so, I've used this global information to populate URLs in my layouts.