Remove limit on number of posts on front page
There are two options:

- Create a fully paginated set of pages with only links to the posts on
  them. Each page would have a limited number of links.

- Show all the posts on the front page.

If I just link to an "all posts" page with no pagination, there's not
much point in limiting the number of posts on the front page. At least,
that will be the case until I redesign the front page to add more
non-post information. For now, just show all the posts on the front
page.