Use server-side generated path for analytics
It turns out that the Pocket app embeds saved webpages into the app, and
when doing so, the embedded browser reports the current pathname
(retrieved via `window.location.pathname`) as a local cached file. To
avoid this, I'm now embedding the path name directly into the Javascript
at the time Jekyll builds the page, avoiding going through the browser.