Style Youtube videos to be sized correctly
I'm still not sure of the exact reason for this, but it seems using
`max-width` on Youtube videos doesn't actually cause a large Youtube
video to take up that max width. Instead, I have to force the video to
that width.

This works because I'm assuming all Youtube videos are large. The reason
for using `max-width` on images instead of `width` is because some
images are skinny enough that we don't want them stretched out the
entire content width. That's not a problem for Youtube videos.