Implement closed B-Spline structure
Uses De Boor's algorithm for evaluating a B-Spline. This structure is
not used yet, but will be used as part of the next article.

Because of the previous commit that separated curve modeling and other
geometric functionality (like sampling, mesh generation, etc.), the
B-Spline structure can simply be slotted into the existing
infrastructure for rendering curves.