Functionality to interpolate twist across curve
This functionality is tied to the Rotation Minimization Frames
functionality, so the twist can only be applied to RMFs. However, the
only reason for this is because the input frames have different
component names depending on how the frames are calculated (i.e. the
Frenet-Serret frames have componest `t`, `n` and `b`, while the RMFs
have components `t`, `r` and `s`). Otherwise, the twist application is
very generic.