CSS: give canvas figures a box shadow
This way, the "non-interactive" figures (which are now sometimes
interactive, in that you can rotate the models), look like the
interactive figures (the ones with form elements below them).

With this change, it looks strange to have the "play button" scale on
hover, as the scale is implemented by scaling the entire canvas. This
wasn't visible before, but now it's extremely visible because the canvas
has the box shadow around it. I just removed the scaling, which I think
is unnecessary.

Probably the only place this looks somewhat out of place is in the first
two WebGL posts, where the affected figures are truly non-interactive.
But even there, it's nice to have a differentiate the live figures from
the static images.