Constrain cover image height on front page
Previously, the cover images would be at least 96px, but would fill up
the entire height of the blog entry. This way, if the blog post title
was really long, the cover image would expand vertically.

Unfortunately, that doesn't seem to play well with a tall cover image,
where, even when the blog post title was short, the cover image expanded
to show the entire image.

I decided having a square preview was generally good no matter what the
size of the blog post title or the cover image.